+++ speclite/filters.py
@@ -11,7 +11,7 @@ with a hyphen, e.g. "sdss2010-r".  The group names included with this package
 are:
 
     >>> filter_group_names
-    ['sdss2010', 'sdss2010noatm', 'decam2014', 'wise2010', 'hsc2017', 'lsst2016', 'bessell', 'BASS', 'MzLS', 'Euclid', 'decamDR1', 'decamDR1noatm', 'gaiadr2', 'twomass']
+    ['sdss2010', 'sdss2010noatm', 'decam2014', 'wise2010', 'hsc2017', 'lsst2016', 'bessell', 'BASS', 'MzLS', 'Euclid', 'decamDR1', 'decamDR1noatm', 'gaiadr2', 'twomass', 'galex', 'odin', 'suprime', 'cfht_megacam']
 
 List the band names associated with any group using, for example:
 
@@ -246,7 +246,8 @@ from .utils import get_path_of_data_file
 
 filter_group_names = [
     'sdss2010', 'sdss2010noatm', 'decam2014', 'wise2010', 'hsc2017', 'lsst2016', 'bessell',
-    'BASS', 'MzLS', 'Euclid', 'decamDR1', 'decamDR1noatm', 'gaiadr2', 'twomass']
+    'BASS', 'MzLS', 'Euclid', 'decamDR1', 'decamDR1noatm', 'gaiadr2', 'twomass', 'galex',
+    'odin', 'suprime', 'cfht_megacam']
 
 default_wavelength_unit = astropy.units.Angstrom
 
@@ -1944,7 +1945,8 @@ def load_filter(name, load_from_cache=True, verbose=False):
 
 def plot_filters(responses, wavelength_unit=None,
                  wavelength_limits=None, wavelength_scale='linear',
-                 legend_loc='upper right', cmap='nipy_spectral'):
+                 legend_loc='upper right', legend_ncols=1, 
+                 response_limits=None, cmap='nipy_spectral'):
     """Plot one or more filter response curves.
 
     The matplotlib package must be installed to use this function. The
@@ -1969,6 +1971,11 @@ def plot_filters(responses, wavelength_unit=None,
     legend_loc : str
         Location of the legend to plot, or do not display any legend if this
         value is None.  See :func:`matplotlib.pyplot.legend` for details.
+    legend_ncols : int
+        Number of legend columns. Default is 1.
+    response_limits : tuple or None
+        Plot limits to use on the response axis, or select limits automatically
+        if this parameter is None.
     cmap : str or :class:`matplotlib.colors.Colormap`
         Color map to use for plotting each filter band.  Colors are assigned
         based on each band's effective wavelength, so a spectral color map
@@ -2022,11 +2029,17 @@ def plot_filters(responses, wavelength_unit=None,
         plt.plot(wlen.value, response.response,
                  color=c, alpha=0.5, label=response.name)
 
-    plt.ylim(0, None)
+    if response_limits is None:
+        plt.ylim(0, None)
+    else:
+        plt.ylim(response_limits)
+
     plt.xlabel('Wavelength [{0}]'.format(wavelength_unit))
     plt.ylabel('Filter Response')
     if legend_loc is not None:
-        plt.legend(loc = legend_loc)
+        if legend_ncols is None:
+            legend_ncols = 1
+        plt.legend(loc=legend_loc, ncol=legend_ncols)
     plt.grid()